Added /api/util.js
