+++ server/api/order.js
@@ -3,6 +3,7 @@ const router = require('express').Router()
 //To get associations, you must import from:
 //this line will navigate us to '../db/models/index.js to get out models.
 const {Order} = require('../db/models')
+const {adminsOnly, adminOrByUserId} = require('./util')
 
 module.exports = router
 
@@ -11,29 +12,6 @@ module.exports = router
 //admin/user may need to see order id, shipping address, order total, order status, and userId associated with the order. We will want admins to be able to edit, add and view the orders.
 
 //Note: a user should only be allowed to see their order, otherwise, seems like we need to limit access
-
-const adminsOnly = (req, res, next) => {
-  //checks to see if there is a logged in user and if yes, then if isAdmin is false for currently logged in user.
-  //Questions: where are we getting user from?
-  if (!req.user || !req.user.isAdmin) {
-    const error = new Error('This page is only viewable by admin users.')
-    //error status response code indicates that the request has not been applied because it lacks valid authentication credentials for the target resource
-    error.status = 401
-    return next(error)
-  }
-  next()
-}
-
-const adminOrByUserId = (req, res, next) => {
-  // This logic should be double checked.
-  if (req.user && (req.user.isAdmin || req.params.id === req.user.id)) {
-    next()
-  } else {
-    const error = new Error('This page is only viewable by admin users.')
-    error.status = 401
-    return next(error)
-  }
-}
 
 //route to view all order information
 //mounted on /api/orders b/c of app.use in index.js in server.
