+++ server/api/users.js
@@ -1,25 +1,7 @@
 const router = require('express').Router()
 const {User} = require('../db/models')
+const {adminsOnly, adminOrByUserId} = require('./util')
 module.exports = router
-
-const adminsOnly = (req, res, next) => {
-  if (!req.user || !req.user.isAdmin) {
-    const error = new Error('This page is only viewable by admin users.')
-    error.status = 401
-    return next(error)
-  }
-  next()
-}
-
-const adminOrByUserId = (req, res, next) => {
-  if (req.user && (req.user.isAdmin || req.params.id === req.user.id)) {
-    next()
-  } else {
-    const error = new Error('This page is only viewable by admin users.')
-    error.status = 401
-    return next(error)
-  }
-}
 
 router.get('/', adminsOnly, async (req, res, next) => {
   try {
